fix library cover icon size type

+++ lib/features/library/widgets/document_card.dart
@@ -145,7 +145,7 @@ class _Cover extends StatelessWidget {
               child: LayoutBuilder(
                 builder: (context, constraints) => Icon(
                   Icons.menu_book_rounded,
-                  size: constraints.biggest.shortestSide.clamp(24, 54),
+                  size: constraints.biggest.shortestSide.clamp(24, 54).toDouble(),
                   color: theme.colorScheme.primary,
                 ),
               ),
